requirements.txt
asgiref==3.6.0
Django==4.1.7
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3

requests~=2.28.2